+++ lib/src/token/token.freezed.dart
@@ -15,7 +15,14 @@ T _$identity<T>(T value) => value;
 /// @nodoc
 mixin _$Token {
 
- String get token;@TimestampConverter() DateTime get validBefore;
+/// [id] is the unique identifier for the token.
+ String? get id;/// [token] is the actual token string that can be used for authentication or other purposes.
+ String get token;/// [validBefore] is the date and time until which the token is valid.
+/// This can be used to determine the token's validity period.
+@TimestampOrNullConverter() DateTime? get validBefore;/// [issuedAt] is the date and time when the token was issued.
+/// This can be used to track when the token was created
+@TimestampOrNullConverter() DateTime? get issuedAt;/// [audience] is the intended audience for the token.
+ TokenAudience? get audience;
 /// Create a copy of Token
 /// with the given fields replaced by the non-null parameter values.
 @JsonKey(includeFromJson: false, includeToJson: false)
@@ -28,16 +35,16 @@ $TokenCopyWith<Token> get copyWith => _$TokenCopyWithImpl<Token>(this as Token, 
 
 @override
 bool operator ==(Object other) {
-  return identical(this, other) || (other.runtimeType == runtimeType&&other is Token&&(identical(other.token, token) || other.token == token)&&(identical(other.validBefore, validBefore) || other.validBefore == validBefore));
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is Token&&(identical(other.id, id) || other.id == id)&&(identical(other.token, token) || other.token == token)&&(identical(other.validBefore, validBefore) || other.validBefore == validBefore)&&(identical(other.issuedAt, issuedAt) || other.issuedAt == issuedAt)&&(identical(other.audience, audience) || other.audience == audience));
 }
 
 @JsonKey(includeFromJson: false, includeToJson: false)
 @override
-int get hashCode => Object.hash(runtimeType,token,validBefore);
+int get hashCode => Object.hash(runtimeType,id,token,validBefore,issuedAt,audience);
 
 @override
 String toString() {
-  return 'Token(token: $token, validBefore: $validBefore)';
+  return 'Token(id: $id, token: $token, validBefore: $validBefore, issuedAt: $issuedAt, audience: $audience)';
 }
 
 
@@ -48,7 +55,7 @@ abstract mixin class $TokenCopyWith<$Res>  {
   factory $TokenCopyWith(Token value, $Res Function(Token) _then) = _$TokenCopyWithImpl;
 @useResult
 $Res call({
- String token,@TimestampConverter() DateTime validBefore
+ String? id, String token,@TimestampOrNullConverter() DateTime? validBefore,@TimestampOrNullConverter() DateTime? issuedAt, TokenAudience? audience
 });
 
 
@@ -65,11 +72,14 @@ class _$TokenCopyWithImpl<$Res>
 
 /// Create a copy of Token
 /// with the given fields replaced by the non-null parameter values.
-@pragma('vm:prefer-inline') @override $Res call({Object? token = null,Object? validBefore = null,}) {
+@pragma('vm:prefer-inline') @override $Res call({Object? id = freezed,Object? token = null,Object? validBefore = freezed,Object? issuedAt = freezed,Object? audience = freezed,}) {
   return _then(_self.copyWith(
-token: null == token ? _self.token : token // ignore: cast_nullable_to_non_nullable
-as String,validBefore: null == validBefore ? _self.validBefore : validBefore // ignore: cast_nullable_to_non_nullable
-as DateTime,
+id: freezed == id ? _self.id : id // ignore: cast_nullable_to_non_nullable
+as String?,token: null == token ? _self.token : token // ignore: cast_nullable_to_non_nullable
+as String,validBefore: freezed == validBefore ? _self.validBefore : validBefore // ignore: cast_nullable_to_non_nullable
+as DateTime?,issuedAt: freezed == issuedAt ? _self.issuedAt : issuedAt // ignore: cast_nullable_to_non_nullable
+as DateTime?,audience: freezed == audience ? _self.audience : audience // ignore: cast_nullable_to_non_nullable
+as TokenAudience?,
   ));
 }
 
@@ -154,10 +164,10 @@ return $default(_that);case _:
 /// }
 /// ```
 
-@optionalTypeArgs TResult maybeWhen<TResult extends Object?>(TResult Function( String token, @TimestampConverter()  DateTime validBefore)?  $default,{required TResult orElse(),}) {final _that = this;
+@optionalTypeArgs TResult maybeWhen<TResult extends Object?>(TResult Function( String? id,  String token, @TimestampOrNullConverter()  DateTime? validBefore, @TimestampOrNullConverter()  DateTime? issuedAt,  TokenAudience? audience)?  $default,{required TResult orElse(),}) {final _that = this;
 switch (_that) {
 case _Token() when $default != null:
-return $default(_that.token,_that.validBefore);case _:
+return $default(_that.id,_that.token,_that.validBefore,_that.issuedAt,_that.audience);case _:
   return orElse();
 
 }
@@ -175,10 +185,10 @@ return $default(_that.token,_that.validBefore);case _:
 /// }
 /// ```
 
-@optionalTypeArgs TResult when<TResult extends Object?>(TResult Function( String token, @TimestampConverter()  DateTime validBefore)  $default,) {final _that = this;
+@optionalTypeArgs TResult when<TResult extends Object?>(TResult Function( String? id,  String token, @TimestampOrNullConverter()  DateTime? validBefore, @TimestampOrNullConverter()  DateTime? issuedAt,  TokenAudience? audience)  $default,) {final _that = this;
 switch (_that) {
 case _Token():
-return $default(_that.token,_that.validBefore);case _:
+return $default(_that.id,_that.token,_that.validBefore,_that.issuedAt,_that.audience);case _:
   throw StateError('Unexpected subclass');
 
 }
@@ -195,10 +205,10 @@ return $default(_that.token,_that.validBefore);case _:
 /// }
 /// ```
 
-@optionalTypeArgs TResult? whenOrNull<TResult extends Object?>(TResult? Function( String token, @TimestampConverter()  DateTime validBefore)?  $default,) {final _that = this;
+@optionalTypeArgs TResult? whenOrNull<TResult extends Object?>(TResult? Function( String? id,  String token, @TimestampOrNullConverter()  DateTime? validBefore, @TimestampOrNullConverter()  DateTime? issuedAt,  TokenAudience? audience)?  $default,) {final _that = this;
 switch (_that) {
 case _Token() when $default != null:
-return $default(_that.token,_that.validBefore);case _:
+return $default(_that.id,_that.token,_that.validBefore,_that.issuedAt,_that.audience);case _:
   return null;
 
 }
@@ -209,12 +219,22 @@ return $default(_that.token,_that.validBefore);case _:
 /// @nodoc
 @JsonSerializable()
 
-class _Token implements Token {
-  const _Token({required this.token, @TimestampConverter() required this.validBefore});
+class _Token extends Token {
+  const _Token({this.id, required this.token, @TimestampOrNullConverter() this.validBefore, @TimestampOrNullConverter() this.issuedAt, this.audience}): super._();
   factory _Token.fromJson(Map<String, dynamic> json) => _$TokenFromJson(json);
 
+/// [id] is the unique identifier for the token.
+@override final  String? id;
+/// [token] is the actual token string that can be used for authentication or other purposes.
 @override final  String token;
-@override@TimestampConverter() final  DateTime validBefore;
+/// [validBefore] is the date and time until which the token is valid.
+/// This can be used to determine the token's validity period.
+@override@TimestampOrNullConverter() final  DateTime? validBefore;
+/// [issuedAt] is the date and time when the token was issued.
+/// This can be used to track when the token was created
+@override@TimestampOrNullConverter() final  DateTime? issuedAt;
+/// [audience] is the intended audience for the token.
+@override final  TokenAudience? audience;
 
 /// Create a copy of Token
 /// with the given fields replaced by the non-null parameter values.
@@ -229,16 +249,16 @@ Map<String, dynamic> toJson() {
 
 @override
 bool operator ==(Object other) {
-  return identical(this, other) || (other.runtimeType == runtimeType&&other is _Token&&(identical(other.token, token) || other.token == token)&&(identical(other.validBefore, validBefore) || other.validBefore == validBefore));
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is _Token&&(identical(other.id, id) || other.id == id)&&(identical(other.token, token) || other.token == token)&&(identical(other.validBefore, validBefore) || other.validBefore == validBefore)&&(identical(other.issuedAt, issuedAt) || other.issuedAt == issuedAt)&&(identical(other.audience, audience) || other.audience == audience));
 }
 
 @JsonKey(includeFromJson: false, includeToJson: false)
 @override
-int get hashCode => Object.hash(runtimeType,token,validBefore);
+int get hashCode => Object.hash(runtimeType,id,token,validBefore,issuedAt,audience);
 
 @override
 String toString() {
-  return 'Token(token: $token, validBefore: $validBefore)';
+  return 'Token(id: $id, token: $token, validBefore: $validBefore, issuedAt: $issuedAt, audience: $audience)';
 }
 
 
@@ -249,7 +269,7 @@ abstract mixin class _$TokenCopyWith<$Res> implements $TokenCopyWith<$Res> {
   factory _$TokenCopyWith(_Token value, $Res Function(_Token) _then) = __$TokenCopyWithImpl;
 @override @useResult
 $Res call({
- String token,@TimestampConverter() DateTime validBefore
+ String? id, String token,@TimestampOrNullConverter() DateTime? validBefore,@TimestampOrNullConverter() DateTime? issuedAt, TokenAudience? audience
 });
 
 
@@ -266,11 +286,14 @@ class __$TokenCopyWithImpl<$Res>
 
 /// Create a copy of Token
 /// with the given fields replaced by the non-null parameter values.
-@override @pragma('vm:prefer-inline') $Res call({Object? token = null,Object? validBefore = null,}) {
+@override @pragma('vm:prefer-inline') $Res call({Object? id = freezed,Object? token = null,Object? validBefore = freezed,Object? issuedAt = freezed,Object? audience = freezed,}) {
   return _then(_Token(
-token: null == token ? _self.token : token // ignore: cast_nullable_to_non_nullable
-as String,validBefore: null == validBefore ? _self.validBefore : validBefore // ignore: cast_nullable_to_non_nullable
-as DateTime,
+id: freezed == id ? _self.id : id // ignore: cast_nullable_to_non_nullable
+as String?,token: null == token ? _self.token : token // ignore: cast_nullable_to_non_nullable
+as String,validBefore: freezed == validBefore ? _self.validBefore : validBefore // ignore: cast_nullable_to_non_nullable
+as DateTime?,issuedAt: freezed == issuedAt ? _self.issuedAt : issuedAt // ignore: cast_nullable_to_non_nullable
+as DateTime?,audience: freezed == audience ? _self.audience : audience // ignore: cast_nullable_to_non_nullable
+as TokenAudience?,
   ));
 }